Rename `ViewPropertiesModal` to `DashboardPropertiesModal`, since we are inly using it for dashboards.
